requirements.txt
alembic==1.12.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlpy==8.0.12
packaging==23.2
protobuf==4.21.12
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
